Migrated the dump and load of test scripts
